feat: refactor authentication classes and update login method signature
