+++ src/main/java/org/mandarin/booking/adapter/persist/MemberQueryRepository.java
@@ -2,7 +2,7 @@ package org.mandarin.booking.adapter.persist;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.mandarin.booking.domain.model.Member;
+import org.mandarin.booking.domain.member.Member;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
