+++ src/main/java/org/mandarin/booking/adapter/webapi/AuthController.java
@@ -2,6 +2,8 @@ package org.mandarin.booking.adapter.webapi;
 
 import jakarta.validation.Valid;
 import org.mandarin.booking.app.port.AuthUseCase;
+import org.mandarin.booking.domain.member.AuthRequest;
+import org.mandarin.booking.domain.member.TokenHolder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,6 +15,6 @@ public record AuthController(AuthUseCase authUsecase) {
 
     @PostMapping("/login")
     public TokenHolder login(@RequestBody @Valid AuthRequest request) {
-        return authUsecase.login(request);
+        return authUsecase.login(request.userId(),request.password());
     }
 }
