+++ src/main/java/org/mandarin/booking/adapter/webapi/MemberController.java
@@ -2,6 +2,8 @@ package org.mandarin.booking.adapter.webapi;
 
 import jakarta.validation.Valid;
 import org.mandarin.booking.app.port.MemberRegisterer;
+import org.mandarin.booking.domain.member.MemberRegisterRequest;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,7 +14,8 @@ import org.springframework.web.bind.annotation.RestController;
 public record MemberController(MemberRegisterer memberRegisterer) {
 
     @PostMapping
-    public MemberRegisterResponse register(@RequestBody @Valid MemberRegisterRequest request){
-        return memberRegisterer.register(request);
+    public ResponseEntity<?> register(@RequestBody @Valid MemberRegisterRequest request){
+        var register = memberRegisterer.register(request);
+        return ResponseEntity.ok(register);
     }
 }
