+++ src/main/java/org/mandarin/booking/app/TokenProvider.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.domain.service;
+package org.mandarin.booking.app;
 
 public interface TokenProvider {
     String generateToken(String userId, String nickName, long expiration);
