+++ src/main/java/org/mandarin/booking/domain/DomainException.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.domain.error;
+package org.mandarin.booking.domain;
 
 public class DomainException extends RuntimeException {
     public DomainException(String message) {
