+++ src/main/java/org/mandarin/booking/domain/member/AuthException.java
@@ -1,4 +1,6 @@
-package org.mandarin.booking.domain.error;
+package org.mandarin.booking.domain.member;
+
+import org.mandarin.booking.domain.DomainException;
 
 public class AuthException extends DomainException {
     public AuthException(String message) {
