+++ src/main/java/org/mandarin/booking/domain/member/Member.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.domain.model;
+package org.mandarin.booking.domain.member;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
@@ -7,8 +7,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import lombok.Getter;
-import org.mandarin.booking.adapter.webapi.MemberRegisterRequest;
-import org.mandarin.booking.domain.service.SecurePasswordEncoder;
+import org.mandarin.booking.app.SecurePasswordEncoder;
 
 @Entity
 @Getter
