+++ src/main/java/org/mandarin/booking/domain/member/MemberException.java
@@ -1,4 +1,6 @@
-package org.mandarin.booking.domain.error;
+package org.mandarin.booking.domain.member;
+
+import org.mandarin.booking.domain.DomainException;
 
 public class MemberException extends DomainException {
     public MemberException(String message) {
