+++ src/main/java/org/mandarin/booking/domain/member/TokenHolder.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.adapter.webapi;
+package org.mandarin.booking.domain.member;
 
 public record TokenHolder(String accessToken, String refreshToken) {
 }
