+++ src/main/java/org/mandarin/booking/domain/member/package-info.java
@@ -1,3 +1,3 @@
 @NonNullApi
-package org.mandarin.booking.domain;
+package org.mandarin.booking.domain.member;
 import org.springframework.lang.NonNullApi;
