+++ src/test/java/org/mandarin/booking/IntegrationTestUtils.java
@@ -3,9 +3,9 @@ package org.mandarin.booking;
 import static org.mandarin.booking.fixture.MemberFixture.EmailGenerator.generateEmail;
 import static org.mandarin.booking.fixture.MemberFixture.NicknameGenerator.generateNickName;
 
-import org.mandarin.booking.adapter.webapi.MemberRegisterRequest;
-import org.mandarin.booking.domain.model.Member;
-import org.mandarin.booking.domain.service.SecurePasswordEncoder;
+import org.mandarin.booking.domain.member.MemberRegisterRequest;
+import org.mandarin.booking.domain.member.Member;
+import org.mandarin.booking.app.SecurePasswordEncoder;
 import org.mandarin.booking.adapter.persist.MemberCommandRepository;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.ResponseEntity;
