+++ src/test/java/org/mandarin/booking/TestConfig.java
@@ -1,6 +1,6 @@
 package org.mandarin.booking;
 
-import org.mandarin.booking.domain.service.SecurePasswordEncoder;
+import org.mandarin.booking.app.SecurePasswordEncoder;
 import org.mandarin.booking.adapter.persist.MemberCommandRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
