+++ src/test/java/org/mandarin/booking/webapi/auth/POST_specs.java
@@ -17,8 +17,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mandarin.booking.IntegrationTest;
 import org.mandarin.booking.IntegrationTestUtils;
-import org.mandarin.booking.adapter.webapi.AuthRequest;
-import org.mandarin.booking.adapter.webapi.TokenHolder;
+import org.mandarin.booking.domain.member.AuthRequest;
+import org.mandarin.booking.domain.member.TokenHolder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.web.client.TestRestTemplate;
